Correção de bug Procedimento widget

+++ lib/presentation/widgets/receita_list_widget.dart
@@ -44,7 +44,7 @@ class ReceitasWidget extends StatelessWidget {
               trailing: IconButton(
                 icon: const Icon(Icons.delete),
                 onPressed: () {
-                  excluirReceita(receita.id as int);
+                  excluirReceita(receita.id);
                 },
               ),
             );
